# cmd/agents-server/internal/docs/swagger.yaml
components:
  schemas:
...
    handler.ErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/protocol.APIError'
      type: object
...
    handler.chatgptCompleteReq:
      properties:
        redirect_url:
          type: string
      type: object
...
    protocol.APIError:
      properties:
        code:
          type: string
        message:
          type: string
      type: object
...
    providers.ChatGPTLoginResult:
      properties:
        authorize_url:
          type: string
      type: object
...
  securitySchemes:
    BearerAuth:
      description: Value is "Bearer <token>" — the token printed at server startup
        (or set via --token).
      in: header
      name: Authorization
      type: apiKey
...
info:
  description: 'REST API of agents-server, the web server over the agents-go SDK:
    sessions, runs, agents, MCP servers, memories, settings, skills, providers, workflows,
    triggers, guardrails, sandboxes and projects. A run starts over REST (POST /sessions/{id}/runs)
    or the WebSocket protocol (docs/reference/protocol.md) and streams over either;
    this spec covers the HTTP surface.'
  title: agents-server API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /providers/{id}/chatgpt/complete:
    post:
      description: Redeems the callback URL the user pastes after authorizing (accepts
        the full URL or its query string).
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/handler.chatgptCompleteReq'
                description: Pasted callback URL
                summary: request
        description: Pasted callback URL
        required: true
      responses:
        "204":
          description: logged in
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Complete ChatGPT login
      tags:
      - providers
  /providers/{id}/chatgpt/login:
    post:
      description: Starts the OAuth flow; open authorize_url in a browser, authorize,
        then submit the resulting callback URL to /chatgpt/complete.
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/providers.ChatGPTLoginResult'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: provider does not use chatgpt_login
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Start ChatGPT login
      tags:
      - providers
  /providers/{id}/chatgpt/logout:
    post:
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "204":
          description: logged out
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.ErrorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: ChatGPT logout
      tags:
      - providers
...
servers:
- url: /api/v1